When comparing community type, cast to u_int16_t. The same is done for the AS. OK henning@

+++ rde_attr.c
@@ -723,7 +723,7 @@ community_match(void *data, u_int16_t len, int as, int type)
 		etype <<= 8;
 		etype |= *p++;
 		if ((as == COMMUNITY_ANY || (u_int16_t)as == eas) &&
-		    (type == COMMUNITY_ANY || type == etype))
+		    (type == COMMUNITY_ANY || (u_int16_t)type == etype))
 			return (1);
 	}
 	return (0);
